+++ storage-reverse-image-search/CHANGELOG.md
@@ -1,3 +1,13 @@
+## Version 0.1.9
+
+fix - bump to nodejs20 runtime in functions and run npm audit fix
+
+fix - if backfill is disabled, create index on first image upload
+
+fix - use a valid default MODEL_URL param
+
+fix - safely get embeddings bucket
+
 ## Version 0.1.8
 
 fix - update default Gemini models